# Conflicts:
#	app/src/main/java/org/schabi/newpipe/player/BackgroundPlayer.java
#	app/src/main/java/org/schabi/newpipe/player/helper/MediaSessionManager.java
#	app/src/main/java/org/schabi/newpipe/util/BitmapUtils.java
#	app/src/main/res/xml/video_audio_settings.xml

+++ app/src/main/java/us/shandian/giga/ui/common/ProgressDrawable.java
@@ -9,6 +9,7 @@ import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
 import android.os.Handler;
 import android.os.Looper;
+
 import androidx.annotation.ColorInt;
 import androidx.annotation.NonNull;
 
@@ -35,8 +36,8 @@ public class ProgressDrawable extends Drawable {
         mForegroundColor = foreground;
     }
 
-    public void setProgress(float progress) {
-        mProgress = progress;
+    public void setProgress(double progress) {
+        mProgress = (float) progress;
         invalidateSelf();
     }
